mw/com: Remove unused functions from ShmPathBuilder
The ShmPathBuilder should be used for creating the path that is provided
to shm_open. It should not contain information about the path that
shm_open prepends (e.g. /dev/shm or /dev/shmem) and also should not have
a separate overload to return just the file name rather than the path.
If this is ever needed, we should instead return a std::path and the
filename can be retrieved from that.

GIT_ORIGIN_SPP_REV_ID: 66a8cd5b57446bef0e8fd2599c80d8c38af7a0c5