bazel: Enable coverage measurements
With this commit, you can generate coverage via:
`bazel coverage //...`

and a coverage report via
`genhtml --branch-coverage --output /tmp/genhtml bazel-out/_coverage/_coverage_report.da`

We introduce coverage support. This way we can measure the structural
coverage achieved for the code. This includes Line and Branch coverage.
It shall be noted that this coverage metrics, right now, only give a
first indication. We are aware of problems within the coverage
calculation.

For example comments are counted as not covered, number of branches is
calculated wrongly.
We want to switch to a safety enable solution as soon as possible, and
are working on to provide the necessary infrastructure.