mw/com: Rename RuntimeBinding to BindingRuntime
We currently are not consistent with the naming of binding runtimes
(i.e. sometimes we use binding runtime and sometimes runtime binding).
This likely comes from the fact that we have proxy bindings, skeleton
bindings etc. However, the binding runtimes are different because the
impl::Runtime does not contain a specific binding which it then
dispatches to. Rather, it contains a map of runtimes for each specific
binding. Therefore, it's clearer to say binding runtime than runtime
binding, since we're referring to the runtime of a given binding.
GIT_ORIGIN_SPP_REV_ID: b9ecdc4af585d1d97f3f01c741b15a120fa8e5ab