lola: Make Skeleton::PrepareOffer not noexcept
Since we want to replace death tests with AMP precondition violation
tests (since the latter are significantly faster than the former), we
need to remove noexcept from the signatures from any functions that we
want to test a precondition violation. Since we will need to create
precondition violation tests for PrepareOffer for testing the addition
of service methods, we remove noexcept from PrepareOffer. We also remove
noexcept from functions which are not in the implementing the
SkeletonBinding interface (changing all functions in the interface is a
larger change which we don't do in this commit).

GIT_ORIGIN_SPP_REV_ID: ae0bfad08a2406e2a429a3f852f99caa4431cd29