lola: Use NiceMocks by default in Skeleton tests
Regular mocks print a message when a function is called on a mock object
which did not explicitly set an expectation for that call. Since we only
want EXPECT_CALLs when the call is the actual expectation being tested in
a test and otherwise use ON_CALL, we expect that there will be a lot of
mocked function calls without explicit expectations. To avoid spamming the
test output, we use NiceMocks by default.

GIT_ORIGIN_SPP_REV_ID: ca1f98b0314664e81dd4baf67f0787e0c442a6ad