lola: Add default behaviour for usage marker file creation in tests
ExpectServiceUsageMarkerFileCreatedOrOpenedAndClosed() now adds explicit
expectations when called. However, the default behaviour for these calls
are added in the constructor of the fixture. This keeps the behaviour
consistent between tests which explicitly check that the marker file was
successfully created and other test which assume this implicitly (i.e.
because of the default behaviours in the fixture constructor).

GIT_ORIGIN_SPP_REV_ID: 6bf9daf3e973168dcbe667def0ed8a8782b7ecbe