lola: Add Proxy PID to ServiceMethodSubscribedHandler
We need the PID of the proxy in the method subscriber handler to
determine whether the handler is being triggered by a proxy which has
restarted. If it has restarted, then the ApplicationId will be the same
but the PID will be different. We need this information to know when to
clean up the old memory regions.

GIT_ORIGIN_SPP_REV_ID: 54cd91fc15fa436474f48132b9d268558c727309